requirements.txt
numpy==1.26.4
pandas==2.2.2
faiss-cpu==1.8.0.post1
sentence_transformers==3.0.1
transformers==4.43.1
gradio==4.38.1
huggingface-hub==0.23.2